construction of testing suite
