+++ Analysis/Stage Testing/SemanticTesting.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Marlowe.Tests
+{
+    [TestClass]
+    internal class SemanticTesting : TestAnalyser{ 
+
+
+    }
+}
