+++ Analysis/Vistior Testing/ClassTests.cs
@@ -1,0 +1,10 @@
+﻿
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Marlowe.Tests{
+    [TestClass]
+    internal class ClassTests : TestAnalyser{ 
+
+
+    }
+}
