+++ Analysis/Vistior Testing/NamespaceTests.cs
@@ -1,0 +1,11 @@
+﻿
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Marlowe.Tests
+{
+    [TestClass]
+    internal class NamespaceTests : TestAnalyser{ 
+
+
+    }
+}
