+++ Analysis/Vistior Testing/OperationalTests.cs
@@ -1,0 +1,10 @@
+﻿
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Marlowe.Tests{
+    [TestClass]
+    internal class OperationalTests : TestAnalyser{ 
+
+
+    }
+}
