+++ Analysis/Vistior Testing/TypeTests.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Marlowe.Tests{
+    [TestClass]
+    internal class TypeTests : TestAnalyser{ 
+
+
+    }
+}
